Move ConstExprEvalHelpers function definitions to .cpp
